feat: split list pages into dedicated components (filters)
