+++ front/src/ui/pages/animals/AnimalDetailPage/components/AnimalBehaviourAccordion.tsx
