+++ front/src/ui/pages/animals/AnimalDetailPage/components/AnimalDeathAccordion.tsx
