+++ front/src/ui/pages/animals/AnimalDetailPage/components/AnimalExitAccordion.tsx
