+++ front/src/ui/pages/animals/AnimalDetailPage/components/AnimalHealthAccordion.tsx
