+++ front/src/ui/pages/animals/AnimalDetailPage/components/AnimalInfoAccordion.tsx
