+++ front/src/ui/pages/animals/AnimalDetailPage/components/AnimalPecAccordion.tsx
