+++ front/src/ui/pages/animals/AnimalDetailPage/components/VeterinarianInterventionModal.tsx
