+++ front/src/ui/pages/animals/AnimalDetailPage/components/VeterinarianInterventionsHistory.tsx
@@ -2,15 +2,15 @@ import React, { FC, useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { MdAddBox, MdAssignment, MdDelete } from "react-icons/md";
 import { Button, Card, CardBody, CardHeader, Col, Row, Table } from "reactstrap";
-import VeterinarianInterventionsManager from "../../../managers/veterinarianInterventions.manager";
+import VeterinarianInterventionsManager from "../../../../../managers/veterinarianInterventions.manager";
 import VeterinarianInterventionModal from "./VeterinarianInterventionModal";
-import VeterinarianIntervention from "../../../logic/entities/VeterinarianIntervention";
+import VeterinarianIntervention from "../../../../../logic/entities/VeterinarianIntervention";
 import toast from "react-hot-toast";
-import Animal from "../../../logic/entities/Animal";
-import useGetPermissions from "../../../hooks/useGetPermissions";
-import { Ressource } from "../../../logic/entities/Permissions";
-import { useVeterinarianInterventionsByAnimal } from "../../../hooks/veterinarianInterventions/useVeterinarianInterventionsByAnimal";
-import { useDeleteVeterinarianIntervention } from "../../../hooks/veterinarianInterventions/useDeleteVeterinarianIntervention";
+import Animal from "../../../../../logic/entities/Animal";
+import useGetPermissions from "../../../../../hooks/useGetPermissions";
+import { Ressource } from "../../../../../logic/entities/Permissions";
+import { useVeterinarianInterventionsByAnimal } from "../../../../../hooks/veterinarianInterventions/useVeterinarianInterventionsByAnimal";
+import { useDeleteVeterinarianIntervention } from "../../../../../hooks/veterinarianInterventions/useDeleteVeterinarianIntervention";
 
 interface VeterinarianInterventionsHistoryProps {
     animal: Animal;
@@ -22,16 +22,16 @@ const VeterinarianInterventionsHistory: FC<VeterinarianInterventionsHistoryProps
     const [modalVeterinarianIntervention, setModalVeterinarianIntervention] = useState<VeterinarianIntervention | null>(null);
 
     const animalId = animal?.id ?? null;
-    const { data: interventionsData, isPending: isInterventionsPending, isError: isInterventionsError, refetch: refetchInterventions } = useVeterinarianInterventionsByAnimal(animalId);
+    const {
+        data: interventionsData,
+        isPending: isInterventionsPending,
+        isError: isInterventionsError,
+        refetch: refetchInterventions,
+    } = useVeterinarianInterventionsByAnimal(animalId);
     const { mutate: deleteInterventionMutation } = useDeleteVeterinarianIntervention();
 
     const veterinarianInterventions = useMemo(
-        () =>
-            interventionsData
-                ? [...interventionsData].sort(
-                      (a, b) => new Date(b.date ?? "").getTime() - new Date(a.date ?? "").getTime()
-                  )
-                : [],
+        () => (interventionsData ? [...interventionsData].sort((a, b) => new Date(b.date ?? "").getTime() - new Date(a.date ?? "").getTime()) : []),
         [interventionsData]
     );
 
@@ -78,9 +78,7 @@ const VeterinarianInterventionsHistory: FC<VeterinarianInterventionsHistoryProps
                                             toast.error(t("animals.message.saveAnimalBeforeIntervention"));
                                             return;
                                         }
-                                        setModalVeterinarianIntervention(
-                                            VeterinarianInterventionsManager.createVeterinarianIntervention()
-                                        );
+                                        setModalVeterinarianIntervention(VeterinarianInterventionsManager.createVeterinarianIntervention());
                                     }}
                                 >
                                     <MdAddBox />
@@ -107,23 +105,15 @@ const VeterinarianInterventionsHistory: FC<VeterinarianInterventionsHistoryProps
                             ) : (
                                 veterinarianInterventions.map((veterinarianIntervention) => (
                                     <tr key={veterinarianIntervention.id}>
-                                        <th scope="row">
-                                            {veterinarianIntervention.dateObject?.readable ?? veterinarianIntervention.date}
-                                        </th>
+                                        <th scope="row">{veterinarianIntervention.dateObject?.readable ?? veterinarianIntervention.date}</th>
                                         <td>{veterinarianIntervention.description}</td>
                                         <td>
-                                            <Button
-                                                color="info"
-                                                onClick={() => showDetail(veterinarianIntervention)}
-                                            >
+                                            <Button color="info" onClick={() => showDetail(veterinarianIntervention)}>
                                                 <MdAssignment />
                                             </Button>
                                         </td>
                                         <td>
-                                            <Button
-                                                color="danger"
-                                                onClick={() => deleteVeterinarianIntervention(veterinarianIntervention)}
-                                            >
+                                            <Button color="danger" onClick={() => deleteVeterinarianIntervention(veterinarianIntervention)}>
                                                 <MdDelete />
                                             </Button>
                                         </td>
