+++ front/src/ui/pages/hostFamilies/HostFamiliesPage.tsx
@@ -1,7 +1,7 @@
 import { FC, useEffect, useMemo, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { Button, Col, Input, Row, Nav, NavItem, NavLink, TabContent, TabPane, Label, Card, CardBody } from "reactstrap";
-import { MdRefresh, MdAssignment, MdAddBox, MdFilterAlt, MdOutlineThumbUp } from "react-icons/md";
+import { Button, Col, Input, Row, Nav, NavItem, NavLink, TabContent, TabPane } from "reactstrap";
+import { MdRefresh, MdAssignment, MdAddBox, MdOutlineThumbUp } from "react-icons/md";
 import { RiZzzFill } from "react-icons/ri";
 import { sortBy } from "../../../utils/sort";
 import { MapContainer, Marker, Popup, TileLayer } from "react-leaflet";
@@ -9,8 +9,6 @@ import L from "leaflet";
 import "leaflet/dist/leaflet.css";
 import { HOST_FAMILY_KIND_ID } from "../../../utils/constants";
 import { BlueIcon, CatIcon, DogIcon, KittenFeedingIcon, KittenIcon, PuppyIcon, UserIcon, NACIcon } from "../../../utils/mapIcons";
-import Switch from "../../components/Switch";
-import Dropdown from "../../components/Dropdown";
 import SortableTable from "../../components/SortableTable";
 import Page, { CustomBreadcrumbItem } from "../../components/Page";
 import HostFamily from "../../../logic/entities/HostFamily";
@@ -20,65 +18,13 @@ import { Ressource } from "../../../logic/entities/Permissions";
 import { useHostFamilies } from "../../../hooks/hostFamilies/useHostFamilies";
 import { useHostFamilyKinds } from "../../../hooks/hostFamilies/useHostFamilyKinds";
 import { useReferents } from "../../../hooks/users/useReferents";
+import HostFamiliesPageFilters, { Filter, FilterType } from "./HostFamiliesPageFilters";
 
 // Leaflet icon fix
 L.Marker.prototype.options.icon = BlueIcon;
 
 interface HostFamiliesPageProps {
     [key: string]: any;
-}
-
-class Filter {
-    value: any;
-    type: FilterType;
-
-    constructor(value: any, type: FilterType) {
-        this.value = value;
-        this.type = type;
-    }
-
-    check(hostFamily: HostFamily): boolean {
-        return FilterType.check(this.type, this.value, hostFamily);
-    }
-}
-
-enum FilterType {
-    MEMBERSHIP_LATE = "En retard de cotisation uniquement",
-    HAS_A_VEHICULE = "Véhiculé·e uniquement",
-    ON_A_BREAK = "Statut",
-    TEMPORARY = "Tampon",
-    REFERENT = "Référent·e",
-    TYPE = "Type de FA",
-}
-
-namespace FilterType {
-    export function isSwitch(filter: FilterType): boolean {
-        switch (filter) {
-            case FilterType.MEMBERSHIP_LATE:
-            case FilterType.HAS_A_VEHICULE:
-                return true;
-            default:
-                return false;
-        }
-    }
-
-    export function check(filter: FilterType, value: any, hostFamily: HostFamily): boolean {
-        if (value === null || value === undefined) return true;
-        switch (filter) {
-            case FilterType.MEMBERSHIP_LATE:
-                return value === true ? hostFamily.membershipUpToDate === false : true;
-            case FilterType.HAS_A_VEHICULE:
-                return value === true ? hostFamily.hasVehicule === value : true;
-            case FilterType.ON_A_BREAK:
-                return hostFamily.onBreak === value;
-            case FilterType.TEMPORARY:
-                return hostFamily.isTemporary === value;
-            case FilterType.REFERENT:
-                return hostFamily.referent?.id === value;
-            case FilterType.TYPE:
-                return hostFamily.hostFamilyKinds?.map((hfk) => hfk.id).includes(value) ?? false;
-        }
-    }
 }
 
 const HostFamiliesPage: FC<HostFamiliesPageProps> = (props) => {
@@ -186,109 +132,6 @@ const HostFamiliesPage: FC<HostFamiliesPageProps> = (props) => {
         return BlueIcon;
     };
 
-    const filterBody = (filter: Filter) => {
-        switch (filter.type) {
-            case FilterType.HAS_A_VEHICULE:
-            case FilterType.MEMBERSHIP_LATE:
-                return (
-                    <Col key={filter.type} className="mb-0">
-                        <Label>
-                            {filter.type === FilterType.MEMBERSHIP_LATE ? t("hostFamilies.filter.membershipLate") : t("hostFamilies.filter.hasVehicule")}
-                        </Label>
-                        <br />
-                        <Switch
-                            disabled={isLoading}
-                            id={filter.type}
-                            isOn={filter.value}
-                            handleToggle={() => {
-                                setFilters((previous) => previous.map((f) => (f.type === filter.type ? new Filter(!f.value, f.type) : f)));
-                            }}
-                        />
-                    </Col>
-                );
-            case FilterType.ON_A_BREAK:
-                return (
-                    <Col key={filter.type} className="mb-0">
-                        <Label>{t("hostFamilies.filter.status")}</Label>
-                        <Dropdown
-                            withNewLine={true}
-                            color={"primary"}
-                            value={filter.value}
-                            values={[true, false, null]}
-                            valueDisplayName={(onBreak: boolean | null) =>
-                                onBreak === null
-                                    ? t("hostFamilies.filter.statusAll")
-                                    : onBreak === true
-                                    ? t("hostFamilies.filter.statusPaused")
-                                    : t("hostFamilies.filter.statusActive")
-                            }
-                            valueActiveCheck={(onBreak: boolean | null) => onBreak === filter.value}
-                            key={"onBreak"}
-                            onChange={(newBreak) => setFilters((previous) => previous.map((f) => (f.type === filter.type ? new Filter(newBreak, f.type) : f)))}
-                        />
-                    </Col>
-                );
-            case FilterType.REFERENT:
-                return (
-                    <Col key={filter.type} className="mb-0">
-                        <Label>{t("hostFamilies.filter.referent")}</Label>
-                        <Dropdown
-                            withNewLine={true}
-                            color={"primary"}
-                            value={referents.find((usr) => usr.id === filter.value)}
-                            values={[...referents, undefined]}
-                            valueDisplayName={(usr) => (usr === undefined ? "-" : `${usr?.name} ${usr?.firstname}`)}
-                            valueActiveCheck={(usr) => usr?.id === filter.value}
-                            key={"referents"}
-                            onChange={(newUser) =>
-                                setFilters((previous) => previous.map((f) => (f.type === filter.type ? new Filter(newUser?.id, f.type) : f)))
-                            }
-                        />
-                    </Col>
-                );
-            case FilterType.TYPE:
-                return (
-                    <Col key={filter.type} className="mb-0">
-                        <Label>{t("hostFamilies.filter.typeFA")}</Label>
-                        <Dropdown
-                            withNewLine={true}
-                            color={"primary"}
-                            value={hostFamilyKinds.find((hfk) => hfk.id === filter.value)}
-                            values={[...hostFamilyKinds, null]}
-                            valueDisplayName={(hfk) => (hfk === null ? "-" : hfk?.name ?? "")}
-                            valueActiveCheck={(hfk) => hfk?.id === filter.value}
-                            key={"hostFamilyKind"}
-                            onChange={(newHFK) => setFilters((previous) => previous.map((f) => (f.type === filter.type ? new Filter(newHFK?.id, f.type) : f)))}
-                        />
-                    </Col>
-                );
-            case FilterType.TEMPORARY:
-                return (
-                    <Col key={filter.type} className="mb-0">
-                        <Label>{t("hostFamilies.filter.temporary")}</Label>
-                        <Dropdown
-                            withNewLine={true}
-                            color={"primary"}
-                            value={filter.value}
-                            values={[true, false, null]}
-                            valueDisplayName={(temporary) =>
-                                temporary === null
-                                    ? t("hostFamilies.filter.statusAll")
-                                    : temporary === true
-                                    ? t("hostFamilies.filter.temporaryYes")
-                                    : t("hostFamilies.filter.temporaryNo")
-                            }
-                            valueActiveCheck={(temporary) => temporary === filter.value}
-                            key={"temporay"}
-                            onChange={(newTemporary) =>
-                                setFilters((previous) => previous.map((f) => (f.type === filter.type ? new Filter(newTemporary, f.type) : f)))
-                            }
-                        />
-                    </Col>
-                );
-        }
-    };
-
     return (
         <Page
             className="HostFamiliesPage"
@@ -320,19 +163,13 @@ const HostFamiliesPage: FC<HostFamiliesPageProps> = (props) => {
                     </Button>
                 </Col>
             </Row>
-            <Card>
-                <CardBody>
-                    <Row>
-                        <Col xs={"auto"} className="mb-0 border-end">
-                            <MdFilterAlt />
-                        </Col>
-                        <Col>
-                            <Row>{filters.filter((f) => FilterType.isSwitch(f.type)).map((filter) => filterBody(filter))}</Row>
-                            <Row>{filters.filter((f) => !FilterType.isSwitch(f.type)).map((filter) => filterBody(filter))}</Row>
-                        </Col>
-                    </Row>
-                </CardBody>
-            </Card>
+            <HostFamiliesPageFilters
+                filters={filters}
+                setFilters={setFilters}
+                hostFamilyKinds={hostFamilyKinds}
+                referents={referents}
+                isLoading={isLoading}
+            />
 
             <br />
