+++ front/src/ui/pages/hostFamilies/HostFamilyDetailPage/components/HostFamilyDetailSummary.tsx
