+++ front/src/ui/pages/hostFamilies/HostFamilyDetailPage/components/HostFamilyHomeAccordion.tsx
